add support for nfss fonts
This is just an experiment. I'm really bad
at user interface so feel free to change it
if you have a better idea.

This adds a function to the patch_font
callback that will patch any font loaded
when patching is allowed (using the \patchttfont
macro), or if it has a name of a font patched before
(to allow different sizes).

It currently won't really work since \select font
is redefined, but it does not interfer. Maybe a package
option to not patch thing should be added to allow this
functionality.